refactor(integrations): fold Settings > Connected accounts into the Integrations page
The global Integrations page is now the one by-app lens: its app detail
dialog carries the per-agent grant toggles that previously lived only in
Settings > Connected accounts. The Settings row survives as a deep link to
the page (same icon, copy, and app count). Deletes the settings section,
ConnectedAppsList, and their now-unused keys; rewrites the IA e2e to
exercise the grant sheet on the page.
