+++ app/src/lib/settings-sections.ts
@@ -6,7 +6,6 @@
  * component module.
  */
 export const SETTINGS_SECTION_IDS = [
-  "connectedAccounts",
   "apiKeys",
   "workspaceContext",
   "userContext",
