+++ app/tests/settings-sections.test.ts
@@ -4,7 +4,7 @@ import { parseSettingsSection } from "../src/lib/settings-sections.ts";
 
 describe("parseSettingsSection", () => {
   it("passes a valid section id through", () => {
-    strictEqual(parseSettingsSection("connectedAccounts"), "connectedAccounts");
+    strictEqual(parseSettingsSection("apiKeys"), "apiKeys");
     strictEqual(parseSettingsSection("reportBug"), "reportBug");
   });
 
@@ -12,6 +12,10 @@ describe("parseSettingsSection", () => {
     strictEqual(parseSettingsSection("nope"), null);
     strictEqual(parseSettingsSection("integrations"), null);
     strictEqual(parseSettingsSection(""), null);
+    // "connectedAccounts" was folded into the global Integrations page (the ONE
+    // by-app lens); the Settings row now deep-links there, so it is no longer a
+    // settings section and a stale deep-link must not land.
+    strictEqual(parseSettingsSection("connectedAccounts"), null);
     // "members" was removed with the Settings > Members surface (the Admin
     // People tab is now the canonical home); a stale deep-link must not land.
     strictEqual(parseSettingsSection("members"), null);
